Refactor CLI plugin services dependencies
